+++ src/main/java/com/example/atp_back/portfolio/PortfolioService.java
@@ -5,9 +5,9 @@ import com.example.atp_back.portfolio.model.entity.*;
 import com.example.atp_back.portfolio.model.request.PortfolioCreateReqDto;
 import com.example.atp_back.portfolio.model.request.PortfolioReplyReq;
 import com.example.atp_back.portfolio.model.response.PortfolioDetailResp;
-import com.example.atp_back.portfolio.model.response.PortfolioInstanceResp;
 import com.example.atp_back.portfolio.model.response.PortfolioListResp;
 import com.example.atp_back.portfolio.model.response.PortfolioPageResp;
+import com.example.atp_back.portfolio.repository.*;
 import com.example.atp_back.user.model.User;
 import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
