+++ src/main/java/com/example/atp_back/portfolio/repository/AcquisitionRepository.java
@@ -1,4 +1,4 @@
-package com.example.atp_back.portfolio;
+package com.example.atp_back.portfolio.repository;
 
 import com.example.atp_back.portfolio.model.entity.Acquisition;
 import org.springframework.data.jpa.repository.JpaRepository;
