+++ src/main/java/com/example/atp_back/portfolio/repository/PortfolioCustomRepositoryImpl.java
@@ -1,4 +1,4 @@
-package com.example.atp_back.portfolio;
+package com.example.atp_back.portfolio.repository;
 
 import com.example.atp_back.portfolio.model.entity.Portfolio;
 import com.querydsl.core.Tuple;
