+++ src/main/java/com/example/atp_back/portfolio/repository/PortfolioReplyLikesRepository.java
@@ -1,4 +1,4 @@
-package com.example.atp_back.portfolio;
+package com.example.atp_back.portfolio.repository;
 
 import com.example.atp_back.portfolio.model.entity.PortfolioReplyLikes;
 import org.springframework.data.jpa.repository.JpaRepository;
